+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - [#11](https://github.com/itsallcode/holiday-calculator/issues/11) Support EOL comments in configuration file.
+- [#8](https://github.com/itsallcode/holiday-calculator/issues/8) Support Orthodox Eastern-based holidays.
 
 ## [0.0.1] - 2021-05-30